Dead Code: level 5 (RISKY)

+++ src/PrestaShopBundle/Form/Admin/Improve/Shipping/Preferences/CarrierOptionsType.php
@@ -54,8 +54,8 @@ class CarrierOptionsType extends TranslatorAwareType
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $carrierChoices = array_merge([
-            'Best price' => (int) -1,
-            'Best grade' => (int) -2,
+            'Best price' => -1,
+            'Best grade' => -2,
         ], $this->carriers);
 
         $builder
